HS-31-02: extract IntelRepository (deferred-intel queue)
Move the 11 intel-queue methods (enqueue/claim/retry/complete/list/summary/
attempts/fail/requeue/update_meeting_intel_status) verbatim out of
MeetingDatabase into holdspeak/db/intel.py. core.py 4311 -> 3934.

requeue_intel_job needs a meeting (cross-domain), so establish the container
back-reference pattern: BaseRepository stores the owning container as self._db
(repos constructed Repo(connection, self)), and requeue calls
self._db.meetings.get_meeting(...). This is the pattern HS-31-03's project
repos will reuse.

19 production + 32 test call sites moved to db.intel.*; 7 fake-db doubles got
an `intel` self-property. intel_snapshots stay with MeetingRepository (embedded
in MeetingState) — this repo owns the jobs/attempts queue only.

Suite green at 2062 passed / 14 skipped (== baseline); db package ruff-clean.
