+++ pm/roadmap/holdspeak/phase-31-database-decomposition/story-02-intel-repository.md
@@ -1,6 +1,8 @@
 # HS-31-02 — `IntelRepository` extract
 
-**Status:** not-started.
+- **Project:** holdspeak
+- **Phase:** 31
+- **Status:** done (2026-06-02). Evidence: [evidence-story-02.md](./evidence-story-02.md).
 
 ## Goal
 
@@ -23,5 +25,5 @@ following the HS-31-01 pattern. Verbatim move, facade delegates.
 
 ## Done when
 
-- [ ] Intel cluster lives in `IntelRepository`; call sites use `db.intel.*`.
-- [ ] Those methods removed from `MeetingDatabase`; full suite green; ruff clean.
+- [x] Intel cluster lives in `IntelRepository`; call sites use `db.intel.*`.
+- [x] Those methods removed from `MeetingDatabase`; full suite green (2062); ruff clean.
